pluginforge/pyproject.toml
[tool.poetry]
name = "pluginforge"
version = "0.1.0"
description = "Application-independent plugin framework for Python"
authors = ["astrapi69"]
license = "MIT"
readme = "README.md"
packages = [{include = "pluginforge"}]

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.24"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
